+++ database/migrations/2020_08_23_175844_create_inventario_piso_ventas_table.php
@@ -17,7 +17,8 @@ class CreateInventarioPisoVentasTable extends Migration
             $table->bigIncrements('id');
             $table->unsignedBigInteger('inventario_id');
             $table->unsignedBigInteger('piso_venta_id');
-            $table->integer('cantidad');
+            //$table->integer('cantidad');
+            $table->decimal('cantidad', 20, 2);
             $table->timestamps();
 
             $table->foreign('inventario_id')->references('id')->on('inventarios');
